fix: 예약 관련 api 폴더구조 수정
