+++ apis/v1/reservation/[id].ts
@@ -1,0 +1,6 @@
+import { getRequest } from '@/apis';
+import { Reservation } from '@/apis/types/reservation';
+
+export const getReservation = async (id: number) => {
+  return await getRequest<Reservation>(`/v1/reservation/${id}.ts`);
+};
